+++ lib/x509.mli
@@ -135,6 +135,10 @@ module Distinguished_name : sig
   (** [pp ppf dn] pretty-prints the distinguished name. *)
   val pp : t Fmt.t
 
+  (** [common_name t] is [Some x] if the distinguished name [t] contains a
+      [CN x], [None] otherwise. *)
+  val common_name : t -> string option
+
   (** [decode_der cs] is [dn], the ASN.1 decoded distinguished name of [cs]. *)
   val decode_der : Cstruct.t -> (t, [> R.msg ]) result